// src/enum/order.enum.ts
export enum EOrderStatus {
	SEARCHING = 'SEARCHING',
	ON_MANAGER = 'ON_MANAGER',
	WAITING = 'WAITING',
	IN_PROGRESS = 'IN_PROGRESS',
	CANCELED = 'CANCELED',
	REQUIRES_PAYMENT = 'REQUIRES_PAYMENT',
	DONE = 'DONE',
}
export enum EResponseStatus {
	WAITING = 'WAITING',
	IN_PATH = 'IN_PATH',
	AWAITING_CONFIRMATION_ARRIVAL = 'AWAITING_CONFIRMATION_ARRIVAL',
	IN_PROGRESS = 'IN_PROGRESS',
	AWAITING_CONFIRMATION_COMPLETION = 'AWAITING_CONFIRMATION_COMPLETION',
	BANNED = 'BANNED',
	DONE = 'DONE',
}